Update touchIt_GET_POS.ino
Known bug - an extra byte is being read by the master, requiring 9 bytes to be read (line 173) and crc check against byte 8 (line 190). Runs smoothly on Arduino but the byte is placed after the data on 5V Arduino and after the length field on the 3.3V Pi setup. Could be master timing or interrupt latency.